evaluation/requirements.txt
# Azure AI Evaluation SDK
azure-ai-evaluation>=1.0.0
azure-identity>=1.15.0
openai>=1.0.0
python-dotenv>=1.0.0
pandas>=2.0.0
aiohttp>=3.9.0
